backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==15.3.4
flake8==6.0.0
mccabe==0.7.0
mysqlclient==2.1.1
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
toposort==1.7
